+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 ### Changes
 
 - Small UX/UI fixes @sneridagh
+- Fix query string search in subjects vocab field @gomez
 - Removed the delete button from the title tile @pnicolli
 
 ### Internal